requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
python-dotenv>=1.0.0
Pillow>=11.0.0
slowapi>=0.1.9
pydantic>=2.0.0
